# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Ascension Quest API
  version: 0.1.0
  description: "Ascension Quest: Legends of Aethoria API"
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /quests:
    get:
      operationId: getQuests
      tags: [quests]
      summary: List quests
      parameters:
        - name: type
          in: query
          schema:
            type: ["string", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Quests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quest"
...
  /quests/{id}:
    get:
      operationId: getQuest
      tags: [quests]
      summary: Get a quest
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quest
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quest"
  /quests/{id}/complete-task:
    post:
      operationId: completeQuestTask
      tags: [quests]
      summary: Mark a quest task as complete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestTaskCompletion"
      responses:
        "200":
          description: Updated quest
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quest"
  /quests/{id}/claim:
    post:
      operationId: claimQuestReward
      tags: [quests]
      summary: Claim quest rewards
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Claim result with rewards
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestClaimResult"

  # ─── INVENTORY ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Player:
      type: object
      required: [id, name, level, rank, xp, xpToNextLevel, hp, maxHp, mp, maxMp, gold, freeStatPoints, stats, activeTitle, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        level:
          type: integer
        rank:
          type: string
          enum: [E, D, C, B, A, S, "National-Level"]
        xp:
          type: integer
        xpToNextLevel:
          type: integer
        hp:
          type: integer
        maxHp:
          type: integer
        mp:
          type: integer
        maxMp:
          type: integer
        gold:
          type: integer
        freeStatPoints:
          type: integer
        streakDays:
          type: integer
        stats:
          $ref: "#/components/schemas/PlayerStats"
        activeTitle:
          type: ["string", "null"]
        penaltyQuestActive:
          type: boolean
        baseClass:
          type: ["string", "null"]
        setupCompleted:
          type: boolean
        createdAt:
          type: string
    PlayerStats:
      type: object
      required: [strength, agility, stamina, vitality, discipline, sense]
      properties:
        strength:
          type: integer
        agility:
          type: integer
        stamina:
          type: integer
        vitality:
          type: integer
        discipline:
          type: integer
        sense:
          type: integer
...
    Quest:
      type: object
      required: [id, title, description, type, status, tasks, xpReward, goldReward, expiresAt, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [daily, weekly, main, side, penalty, boss_raid, gate_dungeon]
        status:
          type: string
          enum: [active, completed, claimed, failed, locked]
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/QuestTask"
        xpReward:
          type: integer
        goldReward:
          type: integer
        bonusStatPoints:
          type: integer
        difficulty:
          type: string
          enum: [E, D, C, B, A, S]
        expiresAt:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        claimedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    QuestTask:
      type: object
      required: [id, questId, description, completed, order]
      properties:
        id:
          type: integer
        questId:
          type: integer
        description:
          type: string
        completed:
          type: boolean
        completedAt:
          type: ["string", "null"]
        targetValue:
          type: ["number", "null"]
        currentValue:
          type: ["number", "null"]
        unit:
          type: ["string", "null"]
        order:
          type: integer
    QuestTaskCompletion:
      type: object
      required: [taskId]
      properties:
        taskId:
          type: integer
        value:
          type: ["number", "null"]
    QuestClaimResult:
      type: object
      required: [xpEarned, goldEarned, newLevel, leveledUp, player]
      properties:
        xpEarned:
          type: integer
        goldEarned:
          type: integer
        bonusStatPoints:
          type: integer
        newLevel:
          type: integer
        leveledUp:
          type: boolean
        newRank:
          type: ["string", "null"]
        rankedUp:
          type: boolean
        player:
          $ref: "#/components/schemas/Player"

    # ─── INVENTORY ─────────────────────────────────────────────────────────────
...